+++ src/opts.rs
@@ -1,55 +1,56 @@
 #[derive(clap::Parser)]
 pub(crate) struct Opt {
-    #[clap(subcommand)]
+    #[command(subcommand)]
     pub command: Option<Command>,
     /// Hide context names in list output
-    #[clap(short = '@')]
+    #[arg(short = '@')]
     pub hide_context: bool,
     /// Hide project names in list output
-    #[clap(short = '+')]
+    #[arg(short = '+')]
     pub hide_project: bool,
     /// Color mode
-    #[clap(short)]
+    #[arg(short)]
     pub color: bool,
     /// Use a configuration file other than one of the defaults
-    #[clap(short = 'd', default_value = "~/.todo/config")]
+    #[arg(short = 'd', default_value = "~/.todo/config")]
     pub config_file: String,
     /// Forces actions without confirmation or interactive input
-    #[clap(short = 'f')]
+    #[arg(short = 'f')]
     pub force: bool,
     /// Plain mode turns off colors
-    #[clap(short)]
+    #[arg(short)]
     pub plain_text: bool,
     /// Hide priority labels in list output.
-    #[clap(short = 'P')]
+    #[arg(short = 'P')]
     pub hide_priority: bool,
     /// Don't auto-archive tasks automatically on completion
-    #[clap(short = 'a')]
+    #[arg(short = 'a')]
     pub dont_auto_archive: bool,
     /// Don't preserve line numbers; automatically remove blank lines on task deletion
-    #[clap(short = 'n')]
+    #[arg(short = 'n')]
     pub dont_preserve_line_numbers: bool,
     /// Prepend the current date to a task automatically when it's added
-    #[clap(short = 't')]
+    #[arg(short = 't')]
     pub append_current_date: bool,
     /// Verbose mode turns on confirmation messages
-    #[clap(short, parse(from_occurrences))]
-    pub verbose: u32,
+    #[arg(short, action = clap::ArgAction::Count)]
+    pub verbose: u8,
     /// Displays version, license and credits
-    #[clap(short = 'V')]
+    #[arg(short = 'V')]
     pub version: bool,
     /// Disables TODOTXT_FINAL_FILTER
-    #[clap(short = 'x')]
+    #[arg(short = 'x')]
     pub disable_final_filter: bool,
 }
 
 #[derive(clap::Subcommand)]
+#[clap(disable_help_subcommand = true)]
 pub(crate) enum Command {
     /// Adds THING I NEED TO DO to your todo.txt file on its own line.
     ///
     /// Project and context notation optional.
     /// Quotes optional.
-    #[clap(alias = "a")]
+    #[command(alias = "a")]
     Add(Add),
 
     /// Adds FIRST THING I NEED TO DO to your todo.txt on its own line and
@@ -68,7 +69,7 @@ pub(crate) enum Command {
     /// Adds TEXT TO APPEND to the end of the task on line ITEM#.
     ///
     /// Quotes optional.
-    #[clap(alias = "app")]
+    #[command(alias = "app")]
     Append(Append),
 
     /// Moves all done tasks from todo.txt to done.txt and removes blank lines.
@@ -80,15 +81,15 @@ pub(crate) enum Command {
     /// Deletes the task on line ITEM# in todo.txt.
     ///
     /// If TERM specified, deletes only TERM from the task.
-    #[clap(alias = "rm")]
+    #[command(alias = "rm")]
     Del(Del),
 
     /// Deprioritizes (removes the priority) from the task(s) on line ITEM# in todo.txt.
-    #[clap(alias = "dp")]
+    #[command(alias = "dp")]
     Delpri(Item),
 
     /// Marks task(s) on line ITEM# as done in todo.txt.
-    #[clap(alias = "do")]
+    #[command(alias = "do")]
     Done(Item),
 
     /// List or flag task ITEM#.
@@ -107,7 +108,7 @@ pub(crate) enum Command {
     /// grep-style basic regular expressions; for literal matching, put a single backslash before
     /// any [ ] \ $ * . ^ and enclose the entire TERM in single quotes, or use double backslashes
     /// and extra shell-quoting.  If no TERM specified, lists entire todo.txt.
-    #[clap(alias = "ls")]
+    #[command(alias = "ls")]
     List(Filter),
 
     /// Displays all the lines in todo.txt AND done.txt that contain TERM(s) sorted by priority
@@ -115,7 +116,7 @@ pub(crate) enum Command {
     ///
     /// Hides all tasks that contain TERM(s) preceded by a minus sign (i.e. -TERM). If no TERM
     /// specified, lists entire todo.txt AND done.txt concatenated and sorted.
-    #[clap(alias = "lsa")]
+    #[command(alias = "lsa")]
     Listall(Filter),
 
     /// Lists all added and overridden actions in the actions directory.
@@ -124,7 +125,7 @@ pub(crate) enum Command {
     /// Lists all the task contexts that start with the @ sign in todo.txt.
     ///
     /// If TERM specified, considers only tasks that contain TERM(s).
-    #[clap(alias = "lsc")]
+    #[command(alias = "lsc")]
     Listcon(Filter),
 
     /// Displays all the lines in SRC file located in the todo.txt directory, sorted by priority
@@ -133,7 +134,7 @@ pub(crate) enum Command {
     /// If TERM specified, lists all lines that contain TERM(s) in SRC file. Hides all tasks that
     /// contain TERM(s) preceded by a minus sign (i.e. -TERM). Without any arguments, the names of
     /// all text files in the todo.txt directory are listed.
-    #[clap(alias = "lf")]
+    #[command(alias = "lf")]
     Listfile(ListFile),
 
     /// Displays all tasks prioritized PRIORITIES.
@@ -141,37 +142,37 @@ pub(crate) enum Command {
     /// PRIORITIES can be a single one (A) or a range (A-C). If no PRIORITIES specified, lists all
     /// prioritized tasks. If TERM specified, lists only prioritized tasks that contain TERM(s).
     /// Hides all tasks that contain TERM(s) preceded by a minus sign (i.e. -TERM).
-    #[clap(alias = "lsp")]
+    #[command(alias = "lsp")]
     Listpri(ListPri),
 
     /// Lists all the projects (terms that start with a + sign) in todo.txt.
     ///
     /// If TERM specified, considers only tasks that contain TERM(s).
-    #[clap(alias = "lsprj")]
+    #[command(alias = "lsprj")]
     Listproj(Filter),
 
     /// Moves a line from source text file (SRC) to destination text file (DEST).
     ///
     /// Both source and destination file must be located in the directory defined in the
     /// configuration directory. When SRC is not defined it's by default todo.txt.
-    #[clap(alias = "mv")]
+    #[command(alias = "mv")]
     Move(Move),
 
     #[cfg(feature = "extended")]
-    #[clap(subcommand)]
+    #[command(subcommand)]
     Note(Note),
 
     /// Adds TEXT TO PREPEND to the beginning of the task on line ITEM#.
     ///
     /// Quotes optional.
-    #[clap(alias = "prep")]
+    #[command(alias = "prep")]
     Prepend(Append),
 
     /// Adds PRIORITY to task on line ITEM#.
     ///
     /// If the task is already prioritized, replaces current priority with new PRIORITY. PRIORITY
     /// must be a letter between A and Z.
-    #[clap(alias = "p")]
+    #[command(alias = "p")]
     Pri(Pri),
 
     /// Replaces task on line ITEM# with UPDATED TODO.
@@ -180,34 +181,34 @@ pub(crate) enum Command {
     /// Adds the number of open tasks and done tasks to report.txt.
     Report,
 
-    #[clap(external_subcommand)]
+    #[command(external_subcommand)]
     External(Vec<String>),
 }
 
 #[derive(clap::Parser)]
 pub(crate) struct Add {
-    #[clap(default_value = "Vec::new")]
+    #[arg(default_value = "Vec::new")]
     pub task: Vec<String>,
 }
 
 #[derive(clap::Parser)]
 pub(crate) struct AddTo {
     pub dest: String,
-    #[clap(flatten)]
+    #[command(flatten)]
     pub add: Add,
 }
 
 #[derive(clap::Parser)]
 pub(crate) struct Append {
     pub item: usize,
-    #[clap(flatten)]
+    #[command(flatten)]
     pub add: Add,
 }
 
 #[derive(clap::Parser)]
 pub(crate) struct Del {
     pub item: usize,
-    #[clap(flatten)]
+    #[command(flatten)]
     pub filter: Filter,
 }
 
@@ -229,14 +230,14 @@ pub(crate) struct Filter {
 #[derive(clap::Parser)]
 pub(crate) struct ListFile {
     pub src: String,
-    #[clap(flatten)]
+    #[command(flatten)]
     pub filter: Filter,
 }
 
 #[derive(clap::Parser)]
 pub(crate) struct ListPri {
     pub priority: Option<char>,
-    #[clap(flatten)]
+    #[command(flatten)]
     pub filter: Filter,
 }
 
@@ -244,7 +245,7 @@ pub(crate) struct ListPri {
 pub(crate) struct Move {
     pub item: usize,
     pub dest: String,
-    #[clap(default_value = "todo.txt")]
+    #[arg(default_value = "todo.txt")]
     pub src: String,
 }
 
@@ -252,11 +253,11 @@ pub(crate) struct Move {
 #[derive(clap::Subcommand)]
 pub(crate) enum Note {
     Archive,
-    #[clap(alias = "a")]
+    #[command(alias = "a")]
     Add(Item),
-    #[clap(alias = "e")]
+    #[command(alias = "e")]
     Edit(Item),
-    #[clap(alias = "s")]
+    #[command(alias = "s")]
     Show(Item),
 }
